paint: increase time resolution

+++ paint/sources/types.h
@@ -214,7 +214,7 @@ typedef struct context {
 	export_destination_t        layers_destination;
 	bool                        export_padding;
 	split_type_t                split_by;
-	f32                         select_time;
+	f64                         select_time;
 	viewport_mode_t             viewport_mode;
 	void                       *viewport_shader;
 	bool                        hscale_was_changed;
@@ -227,7 +227,7 @@ typedef struct context {
 	f32                         last_paint_y;
 	bool                        foreground_event;
 	i32                         painted;
-	f32                         brush_time;
+	f64                         brush_time;
 	f32                         clone_start_x;
 	f32                         clone_start_y;
 	f32                         clone_delta_x;
